@@ -13,25 +13,21 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.9",
-    "@angular/cdk": "^6.4.7",
-    "@angular/common": "^6.1.9",
-    "@angular/compiler": "^6.1.9",
-    "@angular/core": "^6.1.9",
-    "@angular/flex-layout": "^6.0.0-beta.18",
-    "@angular/forms": "^6.1.9",
-    "@angular/http": "^6.1.9",
-    "@angular/material": "^6.0.0",
-    "@angular/platform-browser": "^6.1.9",
-    "@angular/platform-browser-dynamic": "^6.1.9",
-    "@angular/router": "^6.1.9",
-    "angular-pipes": "^8.0.0",
-    "angular-tree-component": "^7.2.0",
+    "@angular/animations": "^8.2.14",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/flex-layout": "^8.0.0-beta.27",
+    "@angular/forms": "^8.2.14",
+    "@angular/material": "^8.2.3",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/router": "^8.2.14",
     "bcrypt": "^3.0.2",
     "body-parser": "^1.18.2",
     "compression": "^1.7.2",
     "cookie-parser": "^1.4.3",
-    "core-js": "2.5.7",
     "cors": "^2.8.4",
     "dotenv": "^6.0.0",
     "events": "^3.0.0",
@@ -47,26 +43,23 @@
     "method-override": "^2.3.10",
     "mongoose": "^5.1.0",
     "morgan": "^1.9.1",
-    "ng2-dragula": "^1.5.0",
     "nodemon": "^1.17.5",
     "passport": "^0.4.0",
     "passport-jwt": "^4.0.0",
     "passport-local": "^1.0.0",
-    "rxjs": "6.2.0",
-    "rxjs-compat": "^6.0.0-rc.0",
+    "rxjs": "^6.5.3",
     "swagger-ui-express": "^3.0.9",
-    "webpack-dev-server": "^3.1.10",
-    "zone.js": "0.8.26"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.7.0-beta.1",
-    "@angular/cli": "^6.1.9",
-    "@angular/compiler-cli": "^6.1.9",
-    "@angular/language-service": "^6.1.9",
+    "@angular-devkit/build-angular": "^0.803.20",
+    "@angular/cli": "^8.3.20",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "^8.2.14",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~10.3.0",
-    "codelyzer": "4.3.0",
+    "@types/node": "~12.12.14",
+    "codelyzer": "^5.2.0",
     "concurrently": "^3.5.1",
     "jasmine-core": "~3.1.0",
     "jasmine-spec-reporter": "~4.2.1",
@@ -75,9 +68,8 @@
     "karma-coverage-istanbul-reporter": "2.0.1",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "1.1.0",
-    "protractor": "^5.4.1",
     "ts-node": "~6.1.0",
-    "tslint": "~5.10.0",
-    "typescript": "~2.9.2"
+    "tslint": "^5.20.1",
+    "typescript": "~3.5.3"
   }
 }
