@@ -3,22 +3,24 @@ import { RouterModule, Routes } from '@angular/router';
 import { AuthGuard } from '../auth/auth-guard.service';
 import { HomeComponent } from '../home/home.component';
 
-const routes: Routes = [{
-  path: '',
-  component: HomeComponent
-}, {
-  path: 'auth',
-  loadChildren: 'app/auth/auth.module#AuthModule'
-}, {
-  path: 'admin',
-  loadChildren: 'app/admin/admin.module#AdminModule'
-}];
+const routes: Routes = [
+  {
+    path: '',
+    component: HomeComponent
+  },
+  {
+    path: 'auth',
+    loadChildren: () => import('../auth/auth.module').then(m => m.AuthModule)
+  },
+  {
+    path: 'admin',
+    loadChildren: () => import('../admin/admin.module').then(m => m.AdminModule)
+  }
+];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
   exports: [RouterModule],
-  providers: [AuthGuard],
-  declarations: []
+  providers: [AuthGuard]
 })
-
 export class AppRoutingModule {}
