@@ -1,12 +1,10 @@
-import { Component, OnInit, OnDestroy } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { Subscription } from 'rxjs/Subscription';
-import { MatIconRegistry } from "@angular/material";
-import { DomSanitizer } from "@angular/platform-browser";
-
+import { Subscription } from 'rxjs';
+import { MatIconRegistry } from '@angular/material';
+import { DomSanitizer } from '@angular/platform-browser';
 
 import { AuthService } from './auth/auth.service';
-import * as schema from './schema/equipment.json';
 
 @Component({
   selector: 'app-root',
@@ -14,7 +12,6 @@ import * as schema from './schema/equipment.json';
   styleUrls: ['./app.component.scss']
 })
 export class AppComponent implements OnInit {
-
   private userSubscription: Subscription;
   public user: any;
 
@@ -24,18 +21,17 @@ export class AppComponent implements OnInit {
     private domSanitizer: DomSanitizer,
     private matIconRegistry: MatIconRegistry
   ) {
-    this.registerSvgIcons()
+    this.registerSvgIcons();
   }
 
   public ngOnInit() {
-
     // init this.user on startup
     this.authService.me().subscribe(data => {
       this.user = data.user;
     });
 
     // update this.user after login/register/logout
-    this.userSubscription = this.authService.$userSource.subscribe((user) => {
+    this.userSubscription = this.authService.$userSource.subscribe(user => {
       this.user = user;
     });
   }
@@ -49,7 +45,7 @@ export class AppComponent implements OnInit {
     this.router.navigate([link]);
   }
 
-  ngOnDestroy() { 
+  ngOnDestroy() {
     if (this.userSubscription) {
       this.userSubscription.unsubscribe();
     }
@@ -88,10 +84,12 @@ export class AppComponent implements OnInit {
       'tow-truck',
       'transportation',
       'trolleybus',
-      'water-transportation',
-    ].forEach((icon) => {
-      this.matIconRegistry.addSvgIcon(icon, this.domSanitizer.bypassSecurityTrustResourceUrl(`assets/icons/${icon}.svg`))
+      'water-transportation'
+    ].forEach(icon => {
+      this.matIconRegistry.addSvgIcon(
+        icon,
+        this.domSanitizer.bypassSecurityTrustResourceUrl(`assets/icons/${icon}.svg`)
+      );
     });
   }
-
 }
