@@ -1,9 +1,4 @@
-/* SystemJS module definition */
-declare var module: NodeModule;
-interface NodeModule {
-  id: string;
-}
-declare module "*.json" {
+declare module '*.json' {
   const value: any;
   export default value;
 }
